Change custom DNS test to use ipconfig instead of resolv.conf
